Move row id reassigner to core
