+++ paimon-core/src/main/java/org/apache/paimon/append/dataevolution/RowRangeMappingIndex.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.flink.dataevolution;
+package org.apache.paimon.append.dataevolution;
 
 import org.apache.paimon.utils.Range;
