+++ paimon-core/src/test/java/org/apache/paimon/append/dataevolution/RowRangeMappingIndexTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.paimon.flink.dataevolution;
+package org.apache.paimon.append.dataevolution;
 
 import org.apache.paimon.utils.Range;
